+++ R/MsBackendMemory-functions.R
@@ -116,3 +116,20 @@ MsBackendMemory <- function() {
     }
     res
 }
+
+.check_peaks_data_value <- function(x, lo) {
+    if (!(is.list(x) || inherits(x, "SimpleList")))
+        stop("'value' has to be a list-like object")
+    if (length(x) != lo)
+        stop("Length of 'value' has to match length of 'object'")
+    if (!(is.matrix(x[[1L]]) | is.data.frame(x[[1L]])))
+        stop("'value' is expected to be a 'list' of 'matrix' ",
+             "or 'data.frame'")
+    cn <- colnames(x[[1L]])
+    lcn <- length(cn)
+    lapply(x, function(z) {
+        cur_cn <- colnames(z)
+        if (lcn != length(cur_cn) || !all(cn == cur_cn))
+            stop("provided matrices don't have the same column names")
+    })
+}